+++ packages/wallet-sdk/src/components/ConnectDialog/ConnectDialog.scss
@@ -1,96 +1,64 @@
-@import "./theme.scss";
+// Copyright (c) 2018-2022 Coinbase, Inc. <https://www.coinbase.com/>
+// Licensed under the Apache License, version 2.0
 
-.sdk-connect-dialog {
-  width: 700px;
-  // height: 430px;
-  margin: auto;
+@import "../theme.scss";
 
-  &-container {
-    border-radius: 12px;
-    padding: 30px;
-
-    &.light {
-      background: $white;
-    }
-
-    &.dark {
-      background: $black;
-    }
-  }
-
-  &-header {
+.-cbwsdk-css-reset {
+  .-cbwsdk-extension-dialog {
+    z-index: 2147483647;
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
     display: flex;
     align-items: center;
-    justify-content: space-between;
-    margin: 0 0 30px;
-  }
+    justify-content: center;
 
-  &-heading {
-    font-style: normal;
-    font-weight: 500;
-    font-size: 28px;
-    line-height: 36px;
-    margin: 0;
-  }
+    $white: #fff;
+    $black: #000;
 
-  &-layout {
-    display: flex;
-    flex-direction: row;
-  }
+    &-backdrop {
+      z-index: 2147483647;
+      position: fixed;
+      top: 0;
+      left: 0;
+      right: 0;
+      bottom: 0;
+      transition: opacity 0.25s;
 
-  &-column-left {
-    margin-right: 30px;
-    flex: 75%;
-  }
+      &.light {
+        background-color: rgba($black, 0.5);
+      }
 
-  &-column-right {
-    flex: 25%;
-    margin-right: 34px;
-  }
-}
+      &.dark {
+        background-color: rgba(50, 53, 61, 0.4);
+      }
 
-.sdk-cancel-button {
-  -webkit-appearance: none;
-  border: none;
-  background: none;
-  cursor: pointer;
-  padding: 0;
-  margin: 0;
-
-  &-x {
-    position: relative;
-    display: block;
-    cursor: pointer;
-
-    &.light {
-      fill: $black;
+      &-hidden {
+        opacity: 0;
+      }
     }
 
-    &.dark {
-      fill: $white;
+    &-box {
+      display: flex;
+      position: relative;
+      flex-direction: column;
+      transform: scale(1);
+      transition: opacity 0.25s, transform 0.25s;
+
+      &-hidden {
+        opacity: 0;
+        transform: scale(0.85);
+      }
     }
-  }
-}
 
-.sdk-wallet-steps {
-  padding: 0 0 0 20px;
-  margin: 0;
-  width: 100%;
+    &-container {
+      display: block;
 
-  &-item {
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    margin-top: 12px;
-  }
-
-  &-item-container {
-    display: flex;
-    align-items: center;
-  }
-
-  &-pad-left {
-    margin-left: 6px;
+      &-hidden {
+        display: none;
+      }
+    }
   }
 }
