--- packages/wallet-sdk/src/components/ConnectDialog/ConnectItem.scss
@@ -1,41 +1,0 @@
-@import "./theme.scss";
-
-.connect-item {
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  padding: 16px 24px;
-  gap: 12px;
-  cursor: pointer;
-
-  &.light {
-    color: $black;
-  }
-
-  &.dark {
-    color: $white;
-  }
-
-  &.selected {
-    background: $blue-wash;
-    border-radius: 100px;
-    color: $blue-dark;
-  }
-
-  &-copy-container {
-    margin-left: 8px;
-  }
-
-  &-title {
-    margin: 0 0 0;
-    font-size: 16px;
-    line-height: 24px;
-  }
-
-  &-description {
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 20px;
-    margin: 0;
-  }
-}
