+++ packages/wallet-sdk/src/components/theme.scss
